All project commented
Disclaimer: I don't believe in commenting code after project is done. This was an exception due to multiple reasons. This is not an excuse, just a disclaimer.
Please comment your code as you code, but not too much because code should be self explanatory (this is arguable and sometimes people understande "self explanatory" as in no comments at all).